CPU_Usage/requirements.txt
influxdb
kafka-python
fastapi
prometheus-client
